refactor: マッピング定義とテーブルをDomain層へ分離
ARKitMapping/SourceMappingクラスと自動マッピングテーブルを
NDMFプラグインのファイルからEditor/Domain/へ移動し、
ドメイン知識をエントリポイントから分離する (#30 Step 1)
