+++ Editor/ARKitBlendShapeGeneratorEditor.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using UnityEngine;
 using UnityEditor;
 using nadena.dev.ndmf.ui;
+using ARKitBlendShapeGenerator.Domain;
 using static ARKitBlendShapeGenerator.Localization;
 
 namespace ARKitBlendShapeGenerator
@@ -964,7 +965,7 @@ namespace ARKitBlendShapeGenerator
             }
 
             var processedAutoNames = new HashSet<string>();
-            foreach (var mapping in BlendShapeProcessor.GetMappingTable())
+            foreach (var mapping in ARKitMappingTable.GetMappings())
             {
                 if (mapping == null || string.IsNullOrEmpty(mapping.arkitName) || mapping.sources == null)
                 {
